requirements.txt
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.1
colorama==0.4.4
cryptography==3.4.8
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
entrypoints==0.3
environ==1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.5.0
parso==0.8.2
phonenumbers==8.12.33
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
Send2Trash==1.8.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
urllib3==1.26.6
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
